강제 형변환 + 연산자

+++ part1/src/part1/OperatorExam.java
@@ -22,7 +22,7 @@ public class OperatorExam {
 		System.out.println(i + j);
 		System.out.println(i - j);
 		System.out.println(i * j);
-		System.out.println(i / j);
+		System.out.println(i / (double)j);
 		System.out.println(i % j);
 	}
 }
